# openbanking-api-docs/swagger/scim-swagger.yaml
...
swagger: '2.0'
info:
  description: These APIs are based on the SCIM 2.0 Protocol (RFC7644). Please note
    that some resource names contain spaces and that may break some Swagger validators.
  version: '2.0'
  title: API Explorer
basePath: '/scim/v2'
...
consumes:
- application/scim+json
produces:
- application/scim+json
paths:
...
  '/OBQualifiedTrustServiceProviders':
    get:
      tags:
      - OBQualifiedTrustServiceProviders Resource Type
      summary: Search for one or more resources
      parameters:
      - '$ref': '#/parameters/attributes'
      - '$ref': '#/parameters/excludedAttributes'
      - '$ref': '#/parameters/filter'
      - '$ref': '#/parameters/sortBy'
      - '$ref': '#/parameters/sortOrder'
      - '$ref': '#/parameters/startIndex'
      - '$ref': '#/parameters/count'
      responses:
        '200':
          description: SCIM 2.0 List Response Message
          schema:
            type: object
            properties:
              Resources:
                type: array
                description: A multi-valued list of complex objects containing the
                  requested resources.
                items:
                  '$ref': '#/definitions/OBQualifiedTrustServiceProviders'
              itemsPerPage:
                type: integer
                format: int32
                description: The number of resources returned in a list response page.
              schemas:
                type: array
                description: Contains a list of one or more URIs that indicate included
                  SCIM schemas that are used to indicate the attributes contained
                  within a resource.
                uniqueItems: true
                items:
                  type: string
                  format: urn
                  enum:
                  - urn:ietf:params:scim:api:messages:2.0:ListResponse
                  default: urn:ietf:params:scim:api:messages:2.0:ListResponse
              startIndex:
                type: integer
                format: int32
                description: The 1-based index of the first result in the current
                  set of list results.
              totalResults:
                type: integer
                format: int32
                description: The total number of results returned by the list or query
                  operation.
        default:
          description: An error occurred while executing the operation
          schema:
            '$ref': '#/definitions/Error Response'
      security:
      - oauth2: []
      externalDocs:
        description: Query Resources
        url: https://tools.ietf.org/html/rfc7644#section-3.4.2
  '/OBQualifiedTrustServiceProviders/.search':
    post:
      tags:
      - OBQualifiedTrustServiceProviders Resource Type
      summary: Search for one or more resources using HTTP POST
      description: Useful for executing searches without passing parameters that may
        contain sensitive information on the URL.
      parameters:
      - in: body
        name: SCIM 2.0 Search Request Message
        required: false
        schema:
          '$ref': '#/definitions/Search Request'
      responses:
        '200':
          description: SCIM 2.0 List Response Message
          schema:
            type: object
            properties:
              Resources:
                type: array
                description: A multi-valued list of complex objects containing the
                  requested resources.
                items:
                  '$ref': '#/definitions/OBQualifiedTrustServiceProviders'
              itemsPerPage:
                type: integer
                format: int32
                description: The number of resources returned in a list response page.
              schemas:
                type: array
                description: Contains a list of one or more URIs that indicate included
                  SCIM schemas that are used to indicate the attributes contained
                  within a resource.
                uniqueItems: true
                items:
                  type: string
                  format: urn
                  enum:
                  - urn:ietf:params:scim:api:messages:2.0:ListResponse
                  default: urn:ietf:params:scim:api:messages:2.0:ListResponse
              startIndex:
                type: integer
                format: int32
                description: The 1-based index of the first result in the current
                  set of list results.
              totalResults:
                type: integer
                format: int32
                description: The total number of results returned by the list or query
                  operation.
        default:
          description: An error occurred while executing the operation
          schema:
            '$ref': '#/definitions/Error Response'
      security:
      - oauth2: []
      externalDocs:
        description: Querying Resources Using HTTP POST
        url: https://tools.ietf.org/html/rfc7644#section-3.4.3
  '/OBQualifiedTrustServiceProviders/{id}':
    get:
      tags:
      - OBQualifiedTrustServiceProviders Resource Type
      summary: Retrieve the resource
      parameters:
      - '$ref': '#/parameters/id'
      - '$ref': '#/parameters/attributes'
      - '$ref': '#/parameters/excludedAttributes'
      responses:
        '200':
          description: The retrieved resource is included in the response body
          schema:
            '$ref': '#/definitions/OBQualifiedTrustServiceProviders'
        default:
          description: An error occurred while executing the operation
          schema:
            '$ref': '#/definitions/Error Response'
      security:
      - oauth2: []
      externalDocs:
        description: Retrieving a Known Resource
        url: https://tools.ietf.org/html/rfc7644#section-3.4.1
...
securityDefinitions:
  oauth2:
    type: oauth2
    authorizationUrl: https://matls-sso.openbanking.org.uk/as/token.oauth2
    flow: implicit
    scopes:
      TPPReadAll: Read access on OBThirdPartyproviders for match on Organisation Id
      ASPSPReadAccess: Read access to  OBAccountPaymentServiceProviders
      TPPReadAccess: Read access on OBThirdPartyproviders
      ASPSPReadAll: Read access on OBAccountPaymentServiceProviders
      QTSPReadAccess: Read access to OBQTSP resource
      AuthoritiesReadAccess: Read access to OBAuthorities resource
definitions:
...
  OBQualifiedTrustServiceProviders:
    required:
    - CertificateFingerprint
    - ServiceCountry
    - ServiceExtensions
    - ServiceLanguage
    - ServiceName
    - ServiceStatus
    - ServiceType
    - StatusStartingTime
    - X509Certificate
    - X509SKI
    - X509SubjectName
    - schemas
    properties:
      CertificateFingerprint:
        type: string
        description: SHA-256 Certificate Fingerprint
        title: CertificateFingerprint
      CreateTimestamp:
        type: string
        description: OB Creation Timestamp
        title: CreateTimestamp
        readOnly: true
      ModifyTimestamp:
        type: string
        description: OB Modification Timestamp
        title: ModifyTimestamp
        readOnly: true
      ServiceCountry:
        type: string
        description: |-
          '<SchemeOperatorAddress>
          <PostalAddresses>
          <PostalAddress xml:lang=''en''>
          <StreetAddress>NG III - Koning Albert II-laan 16</StreetAddress>
          <Locality>Brussels</Locality>
          <StateOrProvince>Brussels</StateOrProvince>
          <PostalCode>1000</PostalCode>
          <CountryName>BE</CountryName>'
        title: ServiceCountry
      ServiceExtensions:
        type: array
        description: '<ServiceInformationExtensions> ... </ServiceInformationExtensions>'
        title: ServiceExtensions
        items:
          type: object
          properties:
            AdditionalInfoLang:
              type: string
              description: |-
                '<AdditionalServiceInformation>
                <URI xml:lang=''en''> http://uri.etsi.org/TrstSvc/TrustedList/SvcInfoExt/ForWebSiteAuthentication </URI>
                </AdditionalServiceInformation>'
              title: AdditionalInfoLang
            AdditionalInfoURI:
              type: string
              description: |-
                '<AdditionalServiceInformation>
                <URI xml:lang=''en''> http://uri.etsi.org/TrstSvc/TrustedList/SvcInfoExt/ForWebSiteAuthentication </URI>
                </AdditionalServiceInformation>'
              title: AdditionalInfoURI
            Critical:
              type: boolean
              description: <Extension Critical='true'>
              title: Critical
      ServiceLanguage:
        type: string
        description: <Name xml:lang='en'>TeleSec ServerPass</Name>
        title: ServiceLanguage
      ServiceName:
        type: string
        description: <Name xml:lang='en'>TeleSec ServerPass</Name>
        title: ServiceName
      ServiceStatus:
        type: string
        description: '<ServiceStatus> http://uri.etsi.org/TrstSvc/TrustedList/Svcstatus/granted</ServiceStatus>'
        title: ServiceStatus
      ServiceType:
        type: string
        title: ServiceType
      StatusStartingTime:
        type: string
        description: '<StatusStartingTime>2017-11-20T09:00:00Z</StatusStartingTime>'
        title: StatusStartingTime
      X509Certificate:
        type: string
        description: '<X509Certificate> MIIF1DCCBLygAwIBAgIIF00 .... </X509Certificate>'
        title: X509Certificate
      X509SKI:
        type: string
        description: '<X509SKI>MCzVT9II0T3jp4KSgjeJbWa5D68=</X509SKI>'
        title: X509SKI
      X509SubjectName:
        type: string
        description: |-
          <X509SubjectName> CN=TeleSec ServerPass Ex ... </X509SubjectName>
          <X509SKI>MCzVT9II0T3jp4KSgjeJbWa5D68=</X509SKI>
        title: X509SubjectName
      externalId:
        type: string
        title: externalId
      id:
        type: string
        title: id
        readOnly: true
      meta:
        type: object
        title: meta
        readOnly: true
        properties:
          created:
            type: string
            format: date-time
            description: Date and time the resource was created
            title: created
            readOnly: true
          lastModified:
            type: string
            format: date-time
            description: Date and time the resource was last modified
            title: lastModified
            readOnly: true
          location:
            type: string
            format: uri
            description: The location (URI) of the resource
            title: location
            readOnly: true
          resourceType:
            type: string
            description: The resource Type
            title: resourceType
            readOnly: true
          version:
            type: string
            description: The version of the resource
            title: version
            readOnly: true
      schemas:
        type: array
        title: schemas
        uniqueItems: true
        items:
          type: string
    description: eIDAS Qualified Trust Service Provider
...
parameters:
  filter:
    name: filter
    in: query
    description: The filter string used to request a subset of resources.
    required: false
    type: string
  startIndex:
    name: startIndex
    in: query
    description: An integer indicating the 1-based index of the first query result.
    required: false
    type: integer
  sortOrder:
    name: sortOrder
    in: query
    description: A string indicating the order in which the 'sortBy' parameter is
      applied.
    required: false
    type: string
    enum:
    - ascending
    - descending
  count:
    name: count
    in: query
    description: An integer indicating the desired maximum number of query results
      per page.
    required: false
    type: integer
  attributes:
    name: attributes
    in: query
    description: A comma separated list of strings indicating the names of resource attributes to return in the response, overriding the set of attributes that would be returned by default. 'excludedAttributes' must not also be specified if this is specified.
    required: false
    type: string
    collectionFormat: csv
  sortBy:
    name: sortBy
    in: query
    description: A string indicating the attribute whose value shall be used to order
      the returned responses.
    required: false
    type: string
  id:
    name: id
    in: path
    description: The resource identifier (value of the 'id' attribute).
    required: true
    type: string
  excludedAttributes:
    name: excludedAttributes
    in: query
    description: A comma separated list of strings indicating the names of resource attributes to be removed from the default set of attributes to return. 'attributes' must not also be specified if this is specified.
    required: false
    type: string
    collectionFormat: csv
...